develop2 fix bugs
