+++ NewsApp/NewsApp/News/SelectedNewsViewController.swift
@@ -1,0 +1,97 @@
+//
+//  SelectedNewsViewController.swift
+//  NewsApp
+//
+//  Created by Азалия Халилова on 01.05.2023.
+//
+
+import UIKit
+import SnapKit
+
+final class SelectedNewsViewController: UIViewController {
+    // MARK: - GUI Variables
+    private lazy var scrollView: UIScrollView = {
+        let scrollView = UIScrollView()
+        scrollView.showsVerticalScrollIndicator = true
+        return scrollView
+    }()
+    
+    private lazy var contentView: UIView = UIView()
+    
+    private lazy var titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "Title"
+        return label
+    }()
+    
+    private lazy var imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.image = UIImage(named: "image")
+        return imageView
+    }()
+    
+    private lazy  var descriptionText: UILabel = {
+        let textView = UILabel()
+        textView.text = "Description. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
+        textView.numberOfLines = 0
+        return textView
+    }()
+    
+    private lazy var timeLabel: UILabel = {
+        let label = UILabel()
+        label.text = "11:11:11"
+        return label
+    }()
+    
+    // MARK: - Life Cycle
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .white
+        navigationController?.navigationBar.prefersLargeTitles = false
+        setupUI()
+    }
+    
+    // MARK: - Private methods
+    
+    private func setupUI() {
+        view.addSubview(scrollView)
+        scrollView.addSubview(contentView)
+        contentView.addSubview(titleLabel)
+        contentView.addSubview(imageView)
+        contentView.addSubview(descriptionText)
+        contentView.addSubview(timeLabel)
+        
+        setupConstraints()
+    }
+    
+    private func setupConstraints() {
+        scrollView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
+            make.bottom.equalTo(view.safeAreaLayoutGuide)
+        }
+        
+        contentView.snp.makeConstraints { make in
+            make.width.edges.equalToSuperview()
+        }
+        
+        titleLabel.snp.makeConstraints { make in
+            make.leading.trailing.top.equalTo(view.safeAreaLayoutGuide).inset(5)
+        }
+        
+        imageView.snp.makeConstraints { make in
+            make.leading.trailing.equalToSuperview().inset(5)
+            make.top.equalTo(titleLabel.snp.bottom).offset(10)
+            make.height.equalTo(view.snp.width)
+        }
+        
+        descriptionText.snp.makeConstraints { make in
+            make.leading.trailing.equalToSuperview().inset(5)
+            make.top.equalTo(imageView.snp.bottom).offset(20)
+        }
+        
+        timeLabel.snp.makeConstraints { make in
+            make.leading.trailing.equalToSuperview().inset(5)
+            make.top.equalTo(descriptionText.snp.bottom).offset(10)
+        }
+    }
+}
